findAddress uses regular c++ types)

+++ src/main.cpp
@@ -7,8 +7,8 @@
 int main(){
     //this stuff should be moved to the initPlayers function
     //player objects should be wrapped in a game class 
-    DWORD64 maxAddr = 0xFFFFFFFF;//Add another F if it starts to fail
-    DWORD p1MacroStateInit = 6482; //initial neutral value for macro state
+    unsigned long long maxAddr = 0xFFFFFFFF;//Add another F if it starts to fail
+    int p1MacroStateInit = 6482; //initial neutral value for macro state
     Player p1;
     Player p2;
     
@@ -20,7 +20,7 @@ int main(){
             Sleep(20);
             initPlayerAddresses(p1,p2);
             //this should be a get function not an assignment
-            p1.macroStateAddr = findAddress<DWORD>(pHandle, p1MacroStateInit, 0x10400264, maxAddr, 0x10000, 4);
+            p1.macroStateAddr = findAddress<int>(pHandle, p1MacroStateInit, 0x10400264, maxAddr, 0x10000, 4);
             if(p1.macroStateAddr){
                 //put these inside the init players function
                 std::cout << "p1 macro state address found\n";
